Optimize claim (and home) lookup
This commit removes most of the costly streams used in the inner loops
when looking up claims. Instead, a simple iterator is utilized.
Additionally, the getClaimAt() function caches its previous result and
attempts to re-use it. The "/ca debug" method will output the amount of
cache hits or misses.

We hope that this alleviates the timings impact of all events;
particularly the (Pre)CreatureSpawnEvent.